backend/requirements.txt
fastapi
uvicorn

pandas
openpyxl

sqlalchemy

llama-cpp-python

python-multipart
